Move submitProjectToGithub to utils
